refactor(brett): index.ts → routes/snapshots+auth+admin+presets (T000661)
